+++ serde_arrow/src/internal/schema/tracer.rs
@@ -525,7 +525,7 @@ fn coerce_primitive_type(
     options: &TracingOptions,
 ) -> Result<(DataType, bool, Option<Strategy>)> {
     use DataType::{
-        Date64, Float32, Float64, Int16, Int32, Int64, Int8, LargeUtf8, Null, UInt16, UInt32,
+        Float32, Float64, Int16, Int32, Int64, Int8, LargeUtf8, Null, Timestamp, UInt16, UInt32,
         UInt64, UInt8, Utf8,
     };
 
@@ -575,11 +575,13 @@ fn coerce_primitive_type(
             (Int8 | Int16 | Int32 | Int64 | UInt8 | UInt16 | UInt32 | UInt64, _),
         ) if options.coerce_numbers => (Float64, nullable, None),
         // incompatible formats, coerce to string
-        ((Date64, nullable, _), (LargeUtf8, _)) => (LargeUtf8, nullable, None),
-        ((LargeUtf8, nullable, _), (Date64, _)) => (LargeUtf8, nullable, None),
-        ((Date64, nullable, _), (Utf8, _)) => (Utf8, nullable, None),
-        ((Utf8, nullable, _), (Date64, _)) => (Utf8, nullable, None),
-        ((Date64, nullable, prev_st), (Date64, curr_st)) if prev_st != curr_st.as_ref() => {
+        ((Timestamp(_, _), nullable, _), (LargeUtf8, _)) => (LargeUtf8, nullable, None),
+        ((LargeUtf8, nullable, _), (Timestamp(_, _), _)) => (LargeUtf8, nullable, None),
+        ((Timestamp(_, _), nullable, _), (Utf8, _)) => (Utf8, nullable, None),
+        ((Utf8, nullable, _), (Timestamp(_, _), _)) => (Utf8, nullable, None),
+        ((Timestamp(_, prev_tz), nullable, _), (Timestamp(_, curr_tz), _))
+            if prev_tz.as_ref() != curr_tz.as_ref() =>
+        {
             (options.string_type(), nullable, None)
         }
         ((prev_ty, _, prev_st), (curr_ty, curr_st)) => {