+++ serde_arrow/src/test/schema_tracing.rs
@@ -171,7 +171,7 @@ mod json_list_null {
 }
 
 /// Strings encoding dates and nulls are traced as nullable Date64 fields
-mod json_date64_naive_null {
+mod json_timestamp_naive_null {
     use super::*;
 
     macro_rules! test {
@@ -181,8 +181,7 @@ mod json_date64_naive_null {
                 let expected = SerdeArrowSchema::from_value(&json!([
                     {
                         "name": "date",
-                        "data_type": "Date64",
-                        "strategy": "NaiveStrAsDate64",
+                        "data_type": "Timestamp(Millisecond, None)",
                         "nullable": true,
                     },
                 ]))?;
@@ -200,7 +199,7 @@ mod json_date64_naive_null {
 }
 
 /// Strings encoding dates and nulls are traced as nullable Date64 fields
-mod json_date64_utc_null {
+mod json_timestamp_utc_null {
     use super::*;
 
     macro_rules! test {
@@ -210,8 +209,7 @@ mod json_date64_utc_null {
                 let expected = SerdeArrowSchema::from_value(&json!([
                     {
                         "name": "date",
-                        "data_type": "Date64",
-                        "strategy": "UtcStrAsDate64",
+                        "data_type": "Timestamp(Millisecond, Some(\"UTC\"))",
                         "nullable": true,
                     },
                 ]))?;